Fix Module ID printf type.

+++ src/app/CableWidget.cpp
@@ -169,14 +169,14 @@ void CableWidget::setCable(engine::Cable* cable) {
 	if (cable) {
 		app::ModuleWidget* outputMw = APP->scene->rack->getModule(cable->outputModule->id);
 		if (!outputMw)
-			throw Exception("Cable cannot find output ModuleWidget %ld", cable->outputModule->id);
+			throw Exception("Cable cannot find output ModuleWidget %lld", (long long) cable->outputModule->id);
 		outputPort = outputMw->getOutput(cable->outputId);
 		if (!outputPort)
 			throw Exception("Cable cannot find output port %d", cable->outputId);
 
 		app::ModuleWidget* inputMw = APP->scene->rack->getModule(cable->inputModule->id);
 		if (!inputMw)
-			throw Exception("Cable cannot find input ModuleWidget %ld", cable->inputModule->id);
+			throw Exception("Cable cannot find input ModuleWidget %lld", (long long) cable->inputModule->id);
 		inputPort = inputMw->getInput(cable->inputId);
 		if (!inputPort)
 			throw Exception("Cable cannot find input port %d", cable->inputId);
